refactor: Separate data layer scope to enable FakeRepository injection
