+++ app/src/main/java/com/metasearch/android/di/AndroidAppGraph.kt
@@ -4,6 +4,7 @@ import android.app.Activity
 import android.content.Context
 import androidx.work.ListenableWorker
 import com.metasearch.android.core.di.ChildWorkerFactory
+import com.metasearch.android.core.di.scope.DataScope
 import dev.zacsweers.metro.AppScope
 import dev.zacsweers.metro.DependencyGraph
 import dev.zacsweers.metro.Multibinds
@@ -11,7 +12,10 @@ import dev.zacsweers.metro.Provider
 import dev.zacsweers.metro.Provides
 import kotlin.reflect.KClass
 
-@DependencyGraph(AppScope::class)
+@DependencyGraph(
+    scope = AppScope::class,
+    additionalScopes = [DataScope::class],
+)
 interface AndroidAppGraph {
 
     @Multibinds
