+++ core/data/impl/src/main/java/com/metasearch/android/core/data/impl/annotation/IoDispatcher.kt
@@ -1,0 +1,7 @@
+package com.metasearch.android.core.data.impl.annotation
+
+import dev.zacsweers.metro.Qualifier
+
+@Qualifier
+@Retention(AnnotationRetention.BINARY)
+annotation class IoDispatcher
