+++ core/data/impl/src/main/java/com/metasearch/android/core/data/impl/repository/DatabaseNameRepositoryImpl.kt
@@ -2,13 +2,14 @@ package com.metasearch.android.core.data.impl.repository
 
 import com.metasearch.android.core.data.api.repository.DatabaseNameRepository
 import com.metasearch.android.core.datastore.api.datasource.DeviceIdDataSource
-import dev.zacsweers.metro.AppScope
+import com.metasearch.android.core.di.scope.DataScope
 import dev.zacsweers.metro.Inject
 import dev.zacsweers.metro.SingleIn
 import java.util.UUID
 
-@SingleIn(AppScope::class)
-class DatabaseNameRepositoryImpl @Inject constructor(
+@SingleIn(DataScope::class)
+@Inject
+class DatabaseNameRepositoryImpl(
     private val deviceIdDataSource: DeviceIdDataSource,
 ) : DatabaseNameRepository {
     override suspend fun getPersistentDeviceDatabaseName(): String {
