+++ core/data/impl/src/main/java/com/metasearch/android/core/data/impl/repository/GraphRepositoryImpl.kt
@@ -4,15 +4,16 @@ import android.net.Uri
 import com.metasearch.android.core.data.api.repository.DatabaseNameRepository
 import com.metasearch.android.core.data.api.repository.GalleryRepository
 import com.metasearch.android.core.data.api.repository.GraphRepository
+import com.metasearch.android.core.di.scope.DataScope
 import com.metasearch.android.core.network.BuildConfig
 import com.metasearch.android.core.network.service.WebService
-import dev.zacsweers.metro.AppScope
 import dev.zacsweers.metro.Inject
 import dev.zacsweers.metro.SingleIn
 import java.net.URLEncoder
 
-@SingleIn(AppScope::class)
-class GraphRepositoryImpl @Inject constructor(
+@SingleIn(DataScope::class)
+@Inject
+class GraphRepositoryImpl(
     private val databaseNameRepository: DatabaseNameRepository,
     private val galleryRepository: GalleryRepository,
     private val webService: WebService,
