+++ core/di/src/main/java/com/metasearch/android/core/di/scope/Scopes.kt
@@ -1,0 +1,4 @@
+package com.metasearch.android.core.di.scope
+
+abstract class DataScope private constructor()
+
